requirements.txt
absl-py==0.5.0
astor==0.7.1
backcall==0.1.0
beautifulsoup4==4.7.0
bleach==3.0.2
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
cloudpickle==0.6.1
cycler==0.10.0
dask==1.0.0
decorator==4.1.2
defusedxml==0.5.0
dlib==19.16.0
dotmap==1.2.20
entrypoints==0.2.3
gast==0.2.0
grpcio==1.15.0
h5py==2.8.0
idna==2.8
ipykernel==4.8.2
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.12.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==2.1.0
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.0
notebook==5.7.0
numpy==1.13.3
olefile==0.44
opencv-contrib-python==3.4.2.16
opencv-python==3.4.2.16
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==4.3.0
prometheus-client==0.4.2
prompt-toolkit==1.0.15
protobuf==3.6.1
ptyprocess==0.6.0
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyWavelets==0.5.2
pyzmq==17.1.2
qtconsole==4.4.2
requests==2.21.0
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.0
seaborn==0.9.0
selenium==3.141.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
soupsieve==1.6.2
tensorboard==1.11.0
tensorboardX==1.4
tensorflow==1.11.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
toolz==0.9.0
torch==0.4.0
torchvision==0.2.1
tornado==4.5.3
tqdm==4.19.4
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
